requirements.txt
Django==3.1.7
djangorestframework==3.12.2
psycopg2-binary>=2.8
python-decouple>=3.4
dj-database-url>=0.5.0
django-cloudinary-storage==0.3.0
requests==2.25.1